All controllers with Authorize

+++ Shop.BLL/Services/ProductService.cs
@@ -22,35 +22,31 @@ namespace Shop.BLL.Services
         public void Add(ProductDTO product)
         {
             var category = unitOfWork.CategoryRepository.GetItemByExpression(n => n.Name.Equals(product.CategoryName, StringComparison.OrdinalIgnoreCase));
-            if (category == null)
-            {
-                throw new NotExsitinException("Category does not existing");
-            }
             var section = unitOfWork.SectionRepository.GetItemByExpression(n => n.Name.Equals(product.SectionName, StringComparison.OrdinalIgnoreCase));
-            if (section == null)
-            {
-                throw new NotExsitinException("Section does not existing");
-            }
             var subsection = unitOfWork.SubSectionRepository.GetItemByExpression(n => n.Name.Equals(product.SubSectionName, StringComparison.OrdinalIgnoreCase));
-            if (subsection == null)
-            {
-                throw new NotExsitinException("Subsection does not existing");
-            }
             Product tempProduct = Mapper.Map<ProductDTO, Product>(product);
-            tempProduct.Category = category;
-            tempProduct.Amount = product.Amount;
-            tempProduct.Section = section;
-            tempProduct.SubSection = subsection;
-            unitOfWork.ProductRepository.Create(tempProduct);
-            unitOfWork.CommtiChanges();
+            var check = unitOfWork.ProductRepository.GetItemByExpression(n => n.Name.Equals(product.Name, StringComparison.OrdinalIgnoreCase));
+            if (check == null)
+            {
+                tempProduct.Category = category ?? throw new NotExsitingException("Category does not existing");
+                tempProduct.Amount = product.Amount;
+                tempProduct.Section = section ?? throw new NotExsitingException("Section does not existing");
+                tempProduct.SubSection = subsection ?? throw new NotExsitingException("Subsection does not existing");
+                unitOfWork.ProductRepository.Create(tempProduct);
+                unitOfWork.CommtiChanges();
+            }
+            else throw new AlreadyExistingException("Product already existing");
         }
 
         public void Delete(int? id)
         {
-
-            unitOfWork.ProductRepository.Delete(id);
-            unitOfWork.CommtiChanges();
-
+            var checkproduct = unitOfWork.ProductCartRepository.GetItemById(id);
+            if (checkproduct != null)
+            {
+                unitOfWork.ProductRepository.Delete(id);
+                unitOfWork.CommtiChanges();
+            }
+            else throw new NotExsitingException("Product Does not existing");
         }
         public void Dispose()
         {
@@ -58,14 +54,13 @@ namespace Shop.BLL.Services
         }
         public ProductDTO GetProduct(int? id)
         {
-            Product product=unitOfWork.ProductRepository.GetItemById(id);
-            return Mapper.Map<Product, ProductDTO>(product);
-        }
-
-        public ProductDTO GetProductByName(string name)
-        {
-            Product product = unitOfWork.ProductRepository.GetItemByExpression(n=>n.Name.Equals(name));
-            return Mapper.Map<Product, ProductDTO>(product);
+            var checkproduct = unitOfWork.ProductCartRepository.GetItemById(id);
+            if (checkproduct != null)
+            {
+                Product product = unitOfWork.ProductRepository.GetItemById(id);
+                return Mapper.Map<Product, ProductDTO>(product);
+            }
+            else throw new NotExsitingException("Product Does not existing");
         }
 
         public IEnumerable<ProductDTO> GetProducts()
@@ -76,15 +71,19 @@ namespace Shop.BLL.Services
 
         public IEnumerable<ProductDTO> GetProductsInRange(int bot, int top)
         {
-            
             IEnumerable<Product> products = unitOfWork.ProductRepository.GetItemListByWhere(p=>p.Pirce>=bot&&p.Pirce<=top);
+            if (products == null)
+                throw new NotExsitingException("Such products not existing");
             var sortedProducts = products.OrderBy(p => p.Pirce);
-            return Mapper.Map<IEnumerable<Product>, List<ProductDTO>>(sortedProducts);
+            var mappedProducts = Mapper.Map<IEnumerable<Product>, List<ProductDTO>>(sortedProducts);
+            return mappedProducts;
         }
         
         public IEnumerable<ProductDTO> GetProductsThatContainsWord(string word)
         {
             IEnumerable<Product> products = unitOfWork.ProductRepository.GetItemListByWhere(n => n.Name.Contains(word));
+            if (products == null)
+                throw new NotExsitingException("this product not existing");
             return Mapper.Map<IEnumerable<Product>, List<ProductDTO>>(products);
 
         }
@@ -94,29 +93,31 @@ namespace Shop.BLL.Services
             var sortedProducts = products.OrderBy(p => p.Pirce);
             return Mapper.Map<IEnumerable<Product>, List<ProductDTO>>(sortedProducts);
         }
-        public void AddToOrder(int? id, int Amount)
-        {
-            var tempProduct = unitOfWork.ProductRepository.GetItemById(id);
-            if (Amount < 0)
-                throw new Exception();
-            if (tempProduct.Amount < Amount)
-                throw new Exception();
-            tempProduct.Amount -= Amount;
-            tempProduct.Sales += Amount;
-            
-        }
+       
 
         public void Update(int? id, ProductDTO product)
         {
             var tempProduct = unitOfWork.ProductRepository.GetItemById(id);
-            var tempCategory= unitOfWork.CategoryRepository.GetItemByExpression(n => n.Name == product.CategoryName);
-            var tempSectiong =unitOfWork.SectionRepository.GetItemByExpression(n => n.Name == product.CategoryName);
-            tempProduct.Amount = product.Amount;
-            tempProduct.Name = product.Name;
-            tempProduct.Pirce = product.Price;
-            tempProduct.Сharacteristics = product.Сharacteristics;
-            tempProduct.Category = tempCategory;
-            tempProduct.CategoryId = tempCategory.Id;
+            if (tempProduct == null)
+                throw new NotExsitingException("this product not existing");
+            var tempCategory= unitOfWork.CategoryRepository.GetItemByExpression(n => n.Name.Equals(product.CategoryName, StringComparison.OrdinalIgnoreCase));
+            var tempSection =unitOfWork.SectionRepository.GetItemByExpression(n => n.Name.Equals(product.SectionName, StringComparison.OrdinalIgnoreCase));
+            var tempSubSection = unitOfWork.SubSectionRepository.GetItemByExpression(n => n.Name.Equals(product.SubSectionName, StringComparison.OrdinalIgnoreCase));
+            var checkDublicateProduct = unitOfWork.ProductRepository.GetItemByExpression(n => n.Name.Equals(product.Name, StringComparison.OrdinalIgnoreCase));
+            if (checkDublicateProduct == null|| tempProduct.Name.Equals(product.Name,StringComparison.OrdinalIgnoreCase))
+            {
+                tempProduct.Amount = product.Amount;
+                tempProduct.Name = product.Name;
+                tempProduct.Pirce = product.Price;
+                tempProduct.Сharacteristics = product.Сharacteristics;
+                tempProduct.Category = tempCategory;
+                tempProduct.CategoryId = tempCategory.Id;
+                tempProduct.Section = tempSection;
+                tempProduct.SubSection = tempSubSection;
+                unitOfWork.ProductRepository.Update(tempProduct);
+                unitOfWork.CommtiChanges();
+            }
+            else throw new AlreadyExistingException("product with this name is already in database");
         }
     }
 }
